requirements.txt
pyyaml
pytelegrambotapi
schedule
python-telegram-bot